==> R/2022/2.0_overview.Rmd <==
---
title: "Landings"
date: "`r Sys.Date()`"
output:
    html_document:
    toc: true
toc_depth: 3
toc_float:
    collapsed: FALSE
code_folding: show
number_sections: TRUE
---
    
    # SET-UP
```{r markdwon, echo=F}
library(knitr)
opts_chunk$set(echo = T, collapse = T, fig.align = "center", fig.width = 9, fig.height = 6)
options(width = 200)
```    

## settings
```{r settings, message = F}
source('0.0_settings.R')
```

# Data

## load
```{r data, message = F}
load(paste0(dir.rdat, "catch.Rdata"))
load(paste0(dir.rdat, "ziff.Rdata"))  # <1995 landings not grouped by DFO region (just province)

gear.groups <- read.csv(paste0(dir.dat,'group_gear_pretty.csv'))
nafo.groups <- read.csv(paste0(dir.dat,'group_nafo_pretty.csv'))
region.groups <- read.csv(paste0(dir.dat,'group_region_pretty.csv'))
TAC <- read.csv(paste0(dir.dat,'TAC.csv'))

catch <- merge(catch, gear.groups,  all.x = T)
catch <- merge(catch, nafo.groups,  all.x = T)
ziff <- merge(ziff, region.groups,  all.x = T)
names(gear.groups)[1] <- 'gear_cat'
ziff <- merge(ziff,gear.groups[,c(1,3)],all.x = T)
```

# Prep for saving
```{r prp, message = F}
y <- max(catch$year,na.rm=T)

wd.img <- paste0('img/',y,'/')
wd.csv <- paste0('csv/',y,'/')

dir.create(wd.img,recursive = T,showWarnings = F)
dir.create(wd.csv,recursive = T,showWarnings = F)
```

# TABLES {.tabset}

## Total
```{r tabel total}
df.tot <- catch
df.tot[df.tot$country!='Canada','country'] <- 'Foreign'
df.tot <- ddply(df.tot,c('year','country'),summarise,catch=sum(catch,na.rm=T))
df.tot <- reshape2::dcast(df.tot,year~country,value.var = 'catch',fill = 0)
df.tot$total <- with(df.tot,Canada+Foreign)

write.csv(df.tot,paste0(wd.csv,'total.csv'),row.names = FALSE)

kable(df.tot,digits=0,row.names = FALSE, caption='Annual landings (t) within Canada’s Exclusive Economic Zone')
```

## Province
```{r tabel province}
df.prov <- catch[catch$source=='ziff',]                                            # how did Thomas got provincial from 1985 onwards?
df.prov <- ddply(df.prov,c('year','prov'),summarise,catch=sum(catch,na.rm=T))
df.prov <- reshape2::dcast(df.prov,year~prov,value.var = 'catch',fill = 0)

write.csv(df.prov,paste0(wd.csv,'byprov.csv'),row.names = FALSE)

kable(df.prov,digits=0,row.names = FALSE, caption='Annual landings (t) by province')
```

## NAFO
```{r table nafo}
df.nafo <- ddply(catch,c('year','nafo'),summarise,catch=sum(catch,na.rm=T))
df.nafo <- reshape2::dcast(df.nafo,year~nafo,value.var = 'catch',fill = 0)

write.csv(df.nafo,paste0(wd.csv,'bynafo.csv'),row.names = FALSE)                                                            

kable(df.nafo,digits=0,row.names = FALSE, caption='Annual landings (t) by NAFO division')
```

## NAFOgroup
```{r table nafogroup}
df.nafogroup <- ddply(catch,c('year','nafogroup'),summarise,catch=sum(catch,na.rm=T))
df.nafogroup <- reshape2::dcast(df.nafogroup,year~nafogroup,value.var = 'catch',fill = 0)

write.csv(df.nafogroup,paste0(wd.csv,'bynafogroup.csv'),row.names = FALSE)                                                            

kable(df.nafogroup,digits=0,row.names = FALSE, caption='Annual landings (t) by NAFO division (grouped)')
```

## DFOregion
```{r table dforeg}
df.dforeg <- ddply(ziff,c('year','region.en'),summarise,catch=sum(catch,na.rm=T)/1000)
df.dforeg <- reshape2::dcast(df.dforeg,year~region.en,value.var = 'catch',fill = 0)

write.csv(df.dforeg,paste0(wd.csv,'bydforeg.csv'),row.names = FALSE)                                                            

kable(df.dforeg,digits=0,row.names = FALSE, caption='Annual landings (t) by DFO region')
```

## Gear (detailed)
```{r table gear}
df.gear <- ddply(catch,c('year','gear'),summarise,catch=sum(catch,na.rm=T))
df.gear <- reshape2::dcast(df.gear,year~gear,value.var = 'catch',fill = 0)

write.csv(df.gear,paste0(wd.csv,'bygear.csv'),row.names = FALSE)

kable(df.gear,digits=0,row.names = FALSE, caption='Annual landings (t) by gear type')

```

## Gear (grouped)
```{r table gear.group}
df.gear <- ddply(catch,c('year','gear.group'),summarise,catch=sum(catch,na.rm=T))
df.gear <- reshape2::dcast(df.gear,year~gear.group,value.var = 'catch',fill = 0)

write.csv(df.gear,paste0(wd.csv,'bygeargroup.csv'),row.names = FALSE)

kable(df.gear,digits=0,row.names = FALSE, caption='Annual landings (t) by gear group')

```

## Target
```{r table target}
df.tar <- ddply(ziff,c('year','species.maintarg_en'),summarise,catch=sum(catch,na.rm=T))
df.tar <- reshape2::dcast(df.tar,year~species.maintarg_en,value.var = 'catch',fill = 0)

write.csv(df.tar,paste0(wd.csv,'bytar.csv'),row.names = FALSE)

kable(df.tar,digits=0,row.names = FALSE, caption='Annual landings (t) by target species')

```

## Gear x DFOregion
```{r table gear.reg}
y <- max(my.year-2): max(my.year)                                                               #(3 last years)
df.gearregion <- ziff[ziff$year %in% y,]
df.gearregion <- ddply(df.gearregion,c('year','gear.group','region.en'),summarise,catch=sum(catch,na.rm=T))
df.gearregion <- reshape2::dcast(df.gearregion,year+region.en~gear.group,value.var = 'catch',fill = 0)

write.csv(df.gearregion,paste0(wd.csv,'bygearRegion_',paste(range(y),collapse = "-"),'.csv'),row.names = FALSE)

kable(df.gearregion,digits=0,row.names = FALSE, caption='Annual landings (kg) by gear type and DFO region')
```

## Top fisheries
```{r table top}
y <- max(my.year-2): max(my.year)  #(3 last years)
# years to consider (one or multiple)
df.top <- catch[catch$year %in% y,]
df.top <- ddply(df.top,c('year','gear','nafo'),summarise,catch=sum(catch,na.rm=T))
df.top <- ddply(df.top,c('year'),transform,catch.perc=catch*100/sum(catch,na.rm=T))
df.top <- ddply(df.top,c('gear','nafo'),summarise,catch.perc=mean(catch.perc,na.rm=T),catch.mean=mean(catch),catch.tot=sum(catch)) # average percentage
df.top <- df.top[order(df.top$catch.perc,decreasing = TRUE),]
df.top$catch.perc <- round(df.top$catch.perc,0)

kable(df.top[df.top$catch.perc>0,],digits=0,row.names = FALSE, caption=paste0('Top fisheries ', max(my.year-2) ,"-", max(my.year)))

```

## Current year
All regions have data (though preliminary)
```{r table year}
y <- max(my.year)            
y
df.2022 <- ziff[ziff$year %in% y & !is.na(ziff$catch),]
df.2022$catch <- df.2022$catch/1000

# total landings (t): 56t
round(sum(df.2022$catch,na.rm = T),0)

# top landings (t): 38t landed by two trips, both targetting mackerel in July (maritimes)
df.2022 <- df.2022[order(df.2022$catch,decreasing = T),]
head(df.2022[,c('catch','species.maincapt_en','species.maintarg_en','land.date','nafo','gear_en','region.en')],10)

# At least from by-catch (but many NA, so don't know): 14t
round(sum(df.2022[df.2022$species.maintarg_en!='Mackerel' &!is.na(df.2022$species.maintarg_en),'catch'],0))

#what by-catch
bycatch <- ddply(df.2022,c('species.maintarg_en'),summarise,catch=sum(catch))
bycatch[order(bycatch$catch,decreasing = TRUE),]

# from vessels with experimental license: nothing
nr <- read.csv(paste0(dir.dat,'npbc_samples.csv'))
df.2022[df.2022$npbc %in% nr[,1],]
```

# PLOTS {.tabset}

## resdoc
Figure for resdoc.
```{r plot all}
# load censor limits
bwd <-  getwd()
setwd('..')                                                                 
load(paste0('./00.0_model/Rdata/',tail(my.year,1),'/input/dat.Rdata'))
ct <- as.data.frame(cbind(dat$aux,exp(dat$logobs)))
ct <- ct[ct$fleet==1,]

# landings because of samples (preliminary for NL)
df.nafogroup <- ddply(catch,c('year','nafogroup'),summarise,catch=sum(catch,na.rm=T))

extrasamp <- data.frame(year=rep(2022,5),
                        nafogroup=c("2GJ3KL",  "4R", "4S", "4T", "4VWX5YZ"),
                        catch=c(5700,3900,2268,3955,3202)/1000)

df.nafogroup <- rbind(df.nafogroup,extrasamp)

p <- ggplot(df.nafogroup,aes(x=year,y=catch/1000))+
    geom_line(data=ct,aes(x=year,y=aux1/1000),col='darkgrey')+
    geom_line(data=ct,aes(x=year,y=aux2/1000),col='darkgrey')+
    geom_bar(stat='identity',aes(fill=nafogroup))+
    geom_point(data=TAC[TAC$type=='CAN',],aes(x=year,y=TAC/1000),size=1)+
    scale_y_continuous(expand=c(0,0),limits=c(0,120))+
    scale_x_continuous(expand=c(0,0))+
    scale_fill_viridis_d(na.value='grey',direction = -1)
#return to project wd
setwd(bwd)   
# plot (English)
pen <- p+labs(y='Landings (kt)',x='Year',fill='Region')
ggsave(paste0('img/',tail(my.year,1),'/catch.png'),pen,width = 15,height = 8,units = 'cm')

#plot (French)
pfr <- p+labs(y='Débarquements (kt)',x='Année',fill='Région')
ggsave(paste0('img/',tail(my.year,1),'/catch_FR.png'),pfr,width = 15,height = 8,units = 'cm')

pen
```

## Gear x DFOregion
```{r table gear.reg3}
y <- max(my.year-2): max(my.year)  #(3 last years)
df.gearregion <- ziff[ziff$year %in% y,]
df.gearregion <- ddply(df.gearregion,c('year','gear.group','region.en'),summarise,catch=sum(catch,na.rm=T))

ggplot(df.gearregion,aes(x=region.en,y=catch/1000,fill=gear.group))+
    geom_bar(stat="identity")+
    facet_grid(year~.,scale="free_y")+
    scale_y_continuous(expand=c(0,0))+
    labs(x='Region',y='Landings (t)',fill='Gear')+
    scale_fill_viridis_d(na.value='grey')

```
