==> R/2022/3.0_comparison.Rmd <==
---
title: "Landings: comparison"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: FALSE
    code_folding: show
    number_sections: TRUE
---

# SET-UP
```{r markdown, echo=F}
library(knitr)
opts_chunk$set(echo = T, collapse = T, fig.align = "center", fig.width = 9, fig.height = 6)
options(width = 200)
```    

## settings
```{r settings, message = F}
source('0.0_settings.R')
```

## load all data
```{r data, message = F}
# New catch
load(paste0(dir.rdat, "catch.Rdata"))

# Old landings
repo <- "https://github.com/iml-mackerel/00.0_model/blob/master/"

ys <- c(2013,2016,2018,2020)
ct.hist <- lapply(ys, function(x) read.ices(url(paste0(repo,'data/',x,'/ct.dat',"?raw=true"))))
names(ct.hist) <- ys 
```

## Binding

```{r wrangle, message = F}
catch.new <- ddply(catch,c('year'),summarise,catch=sum(catch,na.rm = T),source='New')

ctm <- ldply(names(ct.hist), function(x) data.frame(year=as.numeric(row.names(ct.hist[[x]])),catch=ct.hist[[x]][,1],source=x))

catch.comp <- rbind(catch.new,ctm)
```

# PLOT

```{r catch_total_comparison, echo=F}
ggplot(catch.comp[!is.na(catch.comp$catch),],aes(x=year,y=catch,color=source))+
    geom_line(size=1)+
    scale_color_viridis_d()+
    labs(y="Landings (t)",x='Year',color="Source")+
    scale_y_continuous(expand=c(0,0),limits=c(0,max(catch.comp$catch,na.rm = T)*1.05))
```
