==> R/2022/4.0_FAQ.Rmd <==
---
title: "Landings: FAQ"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: FALSE
    code_folding: show
    number_sections: TRUE
---

# SET-UP
```{r markdown, echo=F}
library(knitr)
opts_chunk$set(echo = T, collapse = T, fig.align = "center", fig.width = 9, fig.height = 6)
options(width = 200)
```    

## settings
```{r settings, message = F}
source('0.0_settings.R')
```

## load all data
```{r data, message = F}
load(paste0(dir.rdat, "ziff.Rdata"))
```

# How much do gillnets and handlines/jiggers usually land?
Helpful to determine catch allowed to keep for sample collection.
```{r q1, message = F}
year <- 2011:2021

# subset for years
sub <- ziff[ziff$year %in% year,]

# subset for gears
sub$gear.cat2 <- substr(sub$gear_cat,1,2)
sub <- sub[sub$gear.cat2 %in% c('GN','LH'),]
```



## TABLES AND PLOTS {.tabset}

### Table
```{r table percentile}
sub$c <- round(sub$catch,0)

df <- data.frame(catch=seq(0,1000,100))
df$perc <- round(ecdf(sub$c)(df$catch)*100,0)

kable(df,digits=0,row.names = FALSE, caption='Percentage of landings equal to or lower than indicated')

```

### boxplot overall
```{r plot boxgen}
ggplot(sub,aes(x=gear.cat2,y=catch,fill=as.factor(year)))+
    geom_boxplot()+
    scale_y_log10()+
    labs(y='Landings per trip (kg)',x='Gear categorie',fill='Year')
```

```{r stop}
knitr::knit_exit()
```

### boxplot by region
```{r plot boxreg}
ggplot(sub[sub$prov.land!='NL',],aes(x=gear.cat2,y=catch,fill=as.factor(year)))+
    geom_boxplot()+
    scale_y_log10()+
    labs(y='Landings per trip (kg)',x='Gear categorie',fill='Year')+
    facet_wrap(~prov.land)

```

### ecdf
```{r ecdf}
plot(ecdf(sub$c),xlim=c(0,5000),xlab="Landings per trip (kg)")
```

# Timing of switch from nets to hooks in 4T

## overall
```{r switch to hooks}
lookup <- data.frame(gear.group=c('Gillnet','Driftnet','Handline','Jigger'),
                     hooknet=c('Drift or gillnet','Drift or gillnet','Handline or jigger','Handline or jigger'))
ziffsub <- merge(ziff[ziff$gear.group %in% lookup$gear.group & ziff$nafo=='4T' & !is.na(ziff$gear.group) & !is.na(ziff$nafo),],lookup,all.x=TRUE)
ziffsub$doy <- yday(ziffsub$land.date)

df.hooknet <- ddply(ziffsub,c('hooknet','doy'),summarise,catch=sum(catch,na.rm=T))
ggplot(df.hooknet,aes(x=doy,y=catch,fill=hooknet))+
    geom_rect(aes(xmin=136,xmax=206,ymin=-Inf,ymax=Inf),fill='darkgrey',alpha=0.5)+ # dates from resdoc egg survey Lehoux
    geom_bar(stat='identity')+
    geom_vline(xintercept=171)+
    scale_y_continuous(expand=c(0,0))

```

```{r switch to hooks annual, fig.width=18,fig.height=14, render=F}
df.hooknety <- ddply(ziffsub,c('year','hooknet','doy'),summarise,catch=sum(catch,na.rm=T))
spawning_info <- read_excel("C:/Users/VANBE/Desktop/post-doc/DATA/mack_eggs_biochem/data/spawning_info.xlsx")

ggplot(df.hooknety,aes(x=doy,y=catch,fill=hooknet))+
    geom_rect(data=spawning_info, aes(xmin=136,xmax=206,ymin=-Inf,ymax=Inf),fill='darkgrey',alpha=0.5)+
    geom_bar(stat='identity')+
    geom_vline(xintercept=yday("21-06-2001"))+
    geom_vline(xintercept=171)+
    facet_wrap(~year,scale='free_y')+
    theme(axis.text.y = element_blank())+
    scale_y_continuous(expand=c(0,0))+
    scale_x_continuous(expand=c(0,0),limits=c(80,320))
```
